🧪 datatest2: 약 API 데이터(JSON) 출력 실험2(좌절)
