+++ src/main/java/com/multi/controller/AJAXController.java
@@ -1,8 +1,5 @@
 package com.multi.controller;
 
-import org.json.simple.JSONArray;
-import org.json.simple.JSONObject;
-import org.json.simple.JSONValue;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -14,13 +11,13 @@ import com.multi.restapi.OCREnvelopeAPI;
 /**
  * @author noranbear
  * @date 2022. 7. 6.
- * @version 3.0
+ * @version 4.0
  * @description
  *
  *
- * ===========================================================
+ * =================================================================
  * 	    DATE			 AUTHOR				    NOTE
- * -----------------------------------------------------------
+ * -----------------------------------------------------------------
  *  2022. 7. 6.		    noranbear		     getdata 생성
  *  
  *  2022. 7. 14.				   			evelopescan 생성
@@ -28,7 +25,9 @@ import com.multi.restapi.OCREnvelopeAPI;
  *
  *  2022. 7. 20.	    qwaszx357		dataget controller 수정
  * 
- * ===========================================================
+ *  2022. 7. 21.		noranbear	 dataget test controller 옮기기
+ *  
+ * =================================================================
  */
 @RestController
 
@@ -53,22 +52,7 @@ public class AJAXController {
     @RequestMapping("dataget")
     public Object dataget(String item) throws Exception {
         Object obj = dapi.dataapi(item);
-        System.out.println("result 값 : " + obj);
-        
-        // Object를 JSONObject으로 변환
-        JSONObject jo = (JSONObject) JSONValue.parse(obj.toString());
-        System.out.println("JSONObject로 변환 : " + jo);
-        
-        // jo에서 JSONObject으로 body 뽑아내기
-        JSONObject jo1 = new  JSONObject();
-        jo1 = (JSONObject) jo.get("body");
-        System.out.println("body 뽑아내기 : " + jo1);
-        
-        // body에서 JSONArray로 items 뽑아내기
-        JSONArray ja = new JSONArray();
-        ja = (JSONArray) jo1.get("items");
-        System.out.println("items 뽑아내기 : " + ja);
-        return ja;
+        return obj;
     }
 	
 	/**
