+++ src/main/resources/templates/test/getdataAJAXtest.html
@@ -63,7 +63,7 @@ $(document).ready(function(){
 		//alert(param);
 		$.ajax({
 			method: 'GET',
-			url:'/dataget',
+			url:'/datatestget',
 			data:{'item':param},
 			dataType: 'json',
 			success:function(data){
